removed deprecated optional integer display precision

+++ 12-mysql/1-tuesday-intro/activities/02-FavoriteDB-NoData/Unsolved/favorite_db_unsolved.sql
@@ -15,7 +15,7 @@ CREATE TABLE favorite_songs (
   -- Make a string column called "song" which cannot contain null --
   -- Make a string column called "artist" --
   -- Make an integer column called "score" --
-  score INTEGER(10)
+  score INTEGER
 );
 
 CREATE TABLE favorite_movies (
